+++ src/config.ts
@@ -2,7 +2,7 @@ import { LogLevel, RelativePattern, Uri, workspace, WorkspaceFolder } from "vsco
 import { TestingConfig } from "./types";
 import * as path from "path";
 import lodash from "lodash";
-import { Logger } from "./outputChannel";
+import { Logger } from "./logger";
 
 export namespace ConfigHandler {
     const TESTING_CONFIG_FILE = 'testing.json';
