+++ src/extension.ts
@@ -2,7 +2,7 @@ import { ExtensionContext, LogLevel, workspace } from "vscode";
 import { IBMiTestManager } from "./manager";
 import { getComponentRegistry, getInstance, loadBase } from "./api/ibmi";
 import { Configuration, Section } from "./configuration";
-import { Logger } from "./outputChannel";
+import { Logger } from "./logger";
 import IBMi from "@halcyontech/vscode-ibmi-types/api/IBMi";
 import { RPGUnitComponent } from "./rpgunit";
